Implement Primo cover provider
This is an initial implementation of cover retrieval for Primo.

This includes:
- Basic plumbing for the Primo module and provider setup
- Forms for configuring the Primo integration
- Functionality for retrieving covers from Primo for Ting entities.

This comes with a couple of caveats:
- This is the first stab at implementing Primo integration. It is quite
narrowly focused at the moment. It should be refactored once we begin
implementing Primo search and have a better client library.
- As we do not have actual search results from Primo we fake a set of
known Primo ids which we know will return covers.
- Our development setup currently fails when trying to download covers
from leitir.is. This is probably an issue with https handling. This
needs to be investigated further.

BBS-41