Fix handling of local cover files
Currently ting covers will always consider local cover files unmanaged
and move them - note that there is a typo in the function name.

This is not suitable if the local file is managed as it will be yanked
away from its original reference. In this situation it is more
appropriate to copy it.

Copying will duplicate the file but there is no better way to approach
the situation. ting_covers is deeply tied to covers being available
from a certain path prefix so we cannot reference a file outside of
this. We cannot update the managed file to be located under the ting
covers directory. Ting covers will clean up files here periodically
so this will break the managed file.

Refactor directory setup as it is used for both local and remote files.