Mark which ting object properties are not supported by Primo
We do not support all of the properties included in Ding for Primo.
Mark each of them which we have opted out of and move them to a
separate part of the implementation. This makes them easier to
distinguish and reduces the clutter from the rest of the 
implementation.

BBS-140