Implement search sources provider for Primo
We are not able to extract a list of sources programmatically from
Primo so we provide an administration interface for managing them
manually instead.

The value is stored as an array of sources which can then be used
directly in the provider function.

It must be mapped back and forth from the textarea in the interface.

BBS-161